requirements.txt
Flask
PyPDF2
langchain
google-generativeai
langchain-google-genai
pinecone-client
python-dotenv